Move shim.Command to runtime
